Switched to Tinkoff OpenApi 3

+++ FamilyBudget.v3/App_CodeBase/Tinkoff/Models/IPieDiagramDataItem.cs
@@ -9,8 +9,8 @@ namespace FamilyBudget.v3.App_CodeBase.Tinkoff.Models
     public interface IPieDiagramDataItem
     {
         string Name { get; set; }
-        decimal CurrentTotalInPortfolio { get; set; }
-        decimal CurrentPercentInPortfolio { get; set; }
+        double CurrentTotalInPortfolio { get; set; }
+        double CurrentPercentInPortfolio { get; set; }
         string DiagramBackgroundColor { get; set; }
         string DiagramBackgroundHoverColor { get; set; }
         string DiagramHoverBorderColor { get; set; }
