+++ FamilyBudget.v3/App_CodeBase/Tinkoff/Models/PortfolioDiagramGroupItem.cs
@@ -8,8 +8,8 @@ namespace FamilyBudget.v3.App_CodeBase.Tinkoff.Models
     public class PortfolioDiagramGroupItem : IPieDiagramDataItem
     {
         public string Name { get; set; }
-        public decimal CurrentTotalInPortfolio { get; set; }
-        public decimal CurrentPercentInPortfolio { get; set; }
+        public double CurrentTotalInPortfolio { get; set; }
+        public double CurrentPercentInPortfolio { get; set; }
         public string DiagramBackgroundColor { get; set; }
         public string DiagramBackgroundHoverColor { get; set; }
         public string DiagramHoverBorderColor { get; set; }
